requirements.txt
FastAPI
uvicorn
psycopg2-binary
python-dotenv
vanna
chromadb
google-generativeai
google-cloud-aiplatform